Fix elapsed time converted to float by mistake

+++ data/src/main/kotlin/com/ringoid/data/remote/utils.kt
@@ -107,7 +107,7 @@ inline fun <reified T : BaseResponse> logResponseObservable(tag: String = DEFAUL
 
 // ----------------------------------------------
 fun checkElapsedTimeAndWarn(startTime: Long, tag: String = DEFAULT_TAG) {
-    val elapsedTime = (System.currentTimeMillis() - startTime) * 0.9f
+    val elapsedTime = ((System.currentTimeMillis() - startTime) * 0.9f).toInt()
     if (elapsedTime >= BuildConfig.RESPONSE_TIME_THRESHOLD + 70) {
         val message = "Waiting for $tag longer than expected ${BuildConfig.RESPONSE_TIME_THRESHOLD} ms"; Timber.w(message)
         SentryUtil.w(message, listOf("elapsed time" to "$elapsedTime", "threshold" to "${BuildConfig.RESPONSE_TIME_THRESHOLD}"))
